requirements.txt
pandas
numpy
re
networkx
pyvis.network
IPython.display
IPython.core.display
ast